backend/requirements.txt
Flask
pypdf
reportlab
pytesseract
Pillow
pytest
black
flake8
pre-commit
PyYAML